!24 fix: AdminCategoriesService с категорией не должно быть связано ни одного события.

+++ main/src/main/java/ru/practicum/ewm/endpoints/admin/service/AdminCategoriesServiceImpl.java
@@ -2,17 +2,22 @@ package ru.practicum.ewm.endpoints.admin.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.ewm.dto.categories.CategoryFullDto;
 import ru.practicum.ewm.dto.categories.CategoryInDto;
 import ru.practicum.ewm.exception.CategoryNotFoundException;
 import ru.practicum.ewm.mapper.CategoryMapper;
 import ru.practicum.ewm.model.Category;
 import ru.practicum.ewm.repository.CategoriesRepository;
+import ru.practicum.ewm.repository.EventsRepository;
+
+import java.nio.file.AccessDeniedException;
 
 @Service
 @RequiredArgsConstructor
 public class AdminCategoriesServiceImpl implements AdminCategoriesService {
     private final CategoriesRepository categoriesRepository;
+    private final EventsRepository eventsRepository;
 
     @Override
     public CategoryFullDto updateCategory(CategoryFullDto categoryFullDto) throws CategoryNotFoundException {
@@ -28,8 +33,12 @@ public class AdminCategoriesServiceImpl implements AdminCategoriesService {
     }
 
     @Override
-    public void removeCategory(Long catId) throws CategoryNotFoundException {
+    @Transactional
+    public void removeCategory(Long catId) throws CategoryNotFoundException, AccessDeniedException {
         checkCategoryExists(catId);
+        if (eventsRepository.existsByCategoryId(catId)) {
+            throw new AccessDeniedException("Category in use.");
+        }
         categoriesRepository.deleteById(catId);
     }
